Move color expressions to expressions
