+++ maplibre/expressions.py
@@ -1,5 +1,9 @@
+from __future__ import annotations
+
 import itertools
 from typing import Any, List, TypeVar
+
+from maplibre.colors import color_brewer
 
 T = TypeVar("T")
 
@@ -56,3 +60,30 @@ def quantile_step_expr(
 
     stops = np.quantile(values, probs)
     return step_expr(column, stops, outputs, fallback)
+
+
+def color_quantile_step_expr(
+    column: str, probs: list, values: Any, cmap="viridis"
+) -> list:
+    n = len(probs)
+    colors = color_brewer(cmap, n + 1)
+    return quantile_step_expr(
+        column, probs, outputs=colors[0:n], fallback=colors[-1], values=values
+    )
+
+
+def color_step_expr(column: str, stops: list, cmap="viridis") -> list:
+    n = len(stops)
+    colors = color_brewer(cmap, n + 1)
+    return step_expr(column, stops, outputs=colors[0:n], fallback=colors[-1])
+
+
+def color_match_expr(column: str, categories: Any, cmap: str = "viridis"):
+    categories = sorted(list(set(categories) - {None}))
+    n = len(categories)
+    colors = color_brewer(cmap, n + 1)
+    return match_expr(column, categories, outputs=colors[0:n], fallback=colors[-1])
+
+
+def color_interpolate_linear(column: str, stops: list, cmap: str = "viridis"):
+    pass
